+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The `MultiFetcher` will no longer warn twice for required children without sources.
+- Fix indentation in `enable_verbose_debug_messages()` messages when using `style='rainbow`.  
 
 ## [1.0.2] - 2025-12-03